+++ example_with_tests/lib/fallbacks.dart
@@ -1,0 +1,8 @@
+abstract class Fallbacks {
+  Fallbacks._();
+
+  static const datePattern = 'M/d/yy';
+  static const mediumDatePattern = 'MMM d,y';
+  static const longDatePattern = 'MMMM d,y';
+  static const timePattern = 'h:mm a';
+}