+++ example_with_tests/test/_tools/stubs/stubs.dart
@@ -1,8 +1,16 @@
+import 'package:system_date_time_format/system_date_time_format.dart';
+
 abstract class Stubs {
   Stubs._();
 
-  static const dateFormat = 'dd.MM.y';
-  static const mediumDateFormat = 'd MMM y';
-  static const longDateFormat = 'd MMMM y';
-  static const timeFormat = 'HH:mm';
+  static const datePattern = 'dd.MM.y';
+  static const mediumDatePattern = 'd MMM y';
+  static const longDatePattern = 'd MMMM y';
+  static const timePattern = 'HH:mm';
+  static const allPatterns = Patterns(
+    datePattern: datePattern,
+    mediumDatePattern: mediumDatePattern,
+    longDatePattern: longDatePattern,
+    timePattern: timePattern,
+  );
 }