+++ lib/system_date_time_format.dart
@@ -1,8 +1,15 @@
+import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:system_date_time_format/src/errors/not_ititialized_error.dart';
+import 'package:system_date_time_format/src/errors/sdtf_scope_not_found_error.dart';
 import 'package:system_date_time_format/src/fallbacks.dart';
+import 'package:system_date_time_format/src/patterns.dart';
+import 'package:system_date_time_format/src/widgets/sdtf_scope_inherited.dart';
 
 import 'src/system_date_time_format_platform_interface.dart';
+
+export 'src/patterns.dart';
+export 'src/widgets/sdtf_scope.dart';
 
 /// Provides date and time format from device system settings.
 class SystemDateTimeFormat {
@@ -12,8 +19,56 @@ class SystemDateTimeFormat {
   factory SystemDateTimeFormat() => _instance;
   static final SystemDateTimeFormat _instance = SystemDateTimeFormat._();
 
+  /// Returns a short version of date pattern.
+  /// May throw [PlatformException] from [MethodChannel].
+  Future<String?> getDatePattern() {
+    return SystemDateTimeFormatPlatformInterface.instance.getDatePattern();
+  }
+
+  /// Returns a medium version of date pattern.
+  /// May throw [PlatformException] from [MethodChannel].
+  Future<String?> getMediumDatePattern() {
+    return SystemDateTimeFormatPlatformInterface.instance
+        .getMediumDatePattern();
+  }
+
+  /// Returns a long version of date pattern.
+  /// May throw [PlatformException] from [MethodChannel].
+  Future<String?> getLongDatePattern() {
+    return SystemDateTimeFormatPlatformInterface.instance.getLongDatePattern();
+  }
+
+  /// Returns time pattern.
+  /// May throw [PlatformException] from [MethodChannel].
+  Future<String?> getTimePattern() {
+    return SystemDateTimeFormatPlatformInterface.instance.getTimePattern();
+  }
+
+  /// Returns all available date & time patterns.
+  /// May throw [PlatformException] from [MethodChannel].
+  Future<Patterns> getAllPatterns() async {
+    return Patterns(
+      datePattern: await getDatePattern(),
+      mediumDatePattern: await getMediumDatePattern(),
+      longDatePattern: await getLongDatePattern(),
+      timePattern: await getTimePattern(),
+    );
+  }
+
+  static Patterns of(BuildContext context) {
+    final sDTFScope =
+        context.dependOnInheritedWidgetOfExactType<SDTFScopeInherited>();
+    if (sDTFScope == null) {
+      throw SDTFScopeNotFoundError(context.widget.runtimeType);
+    }
+
+    return sDTFScope.patterns;
+  }
+
   /// Initializes the plugin.
   /// Call this method before using the plugin further.
+  /// It will be removed in 1.0.0
+  @Deprecated("Use SDTFScope instead.")
   Future<void> initialize({
     String dateFormatFallback = Fallbacks.dateFormat,
     String mediumDateFormatFallback = Fallbacks.mediumDateFormat,
@@ -22,25 +77,27 @@ class SystemDateTimeFormat {
   }) async {
     final platform = SystemDateTimeFormatPlatformInterface.instance;
     _dateFormat = await _try(
-      platform.getDateFormat,
+      platform.getDatePattern,
       fallback: dateFormatFallback,
     );
     _mediumDateFormat = await _try(
-      platform.getMediumDateFormat,
+      platform.getMediumDatePattern,
       fallback: mediumDateFormatFallback,
     );
     _longDateFormat = await _try(
-      platform.getLongDateFormat,
+      platform.getLongDatePattern,
       fallback: longDateFormatFallback,
     );
     _timeFormat = await _try(
-      platform.getTimeFormat,
+      platform.getTimePattern,
       fallback: timeFormatFallback,
     );
   }
 
   /// Returns a short version of date format.
   /// Throws [NotInitializedError] when plugin was not initialized
+  /// It will be removed in 1.0.0
+  @Deprecated("Use getDatePattern() instead.")
   String get dateFormat {
     if (_dateFormat == null) {
       throw NotInitializedError('dateFormat');
@@ -50,6 +107,8 @@ class SystemDateTimeFormat {
 
   /// Returns a medium version of date format.
   /// Throws [NotInitializedError] when plugin was not initialized
+  /// It will be removed in 1.0.0
+  @Deprecated("Use getMediumDatePattern() instead.")
   String get mediumDateFormat {
     if (_mediumDateFormat == null) {
       throw NotInitializedError('mediumDateFormat');
@@ -59,6 +118,8 @@ class SystemDateTimeFormat {
 
   /// Returns a long version of date format.
   /// Throws [NotInitializedError] when plugin was not initialized
+  /// It will be removed in 1.0.0
+  @Deprecated("Use getLongDatePattern() instead.")
   String get longDateFormat {
     if (_longDateFormat == null) {
       throw NotInitializedError('longDateFormat');
@@ -68,6 +129,8 @@ class SystemDateTimeFormat {
 
   /// Returns time format.
   /// Throws [NotInitializedError] when plugin was not initialized
+  /// It will be removed in 1.0.0
+  @Deprecated("Use getTimePattern() instead.")
   String get timeFormat {
     if (_timeFormat == null) {
       throw NotInitializedError('timeFormat');
@@ -77,11 +140,11 @@ class SystemDateTimeFormat {
 
   /// Helper function for catching [PlatformException] and returning a fallback when that happen
   Future<String> _try(
-    Future<String> Function({required String fallback}) function, {
+    Future<String?> Function() function, {
     required String fallback,
   }) async {
     try {
-      return await function(fallback: fallback);
+      return await function() ?? fallback;
     } on PlatformException {
       return fallback;
     }