+++ test/_tools/mocks/mocks.dart
@@ -1,5 +1,7 @@
 import 'package:mocktail/mocktail.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 import 'package:system_date_time_format/src/system_date_time_format_platform_interface.dart';
+import 'package:system_date_time_format/system_date_time_format.dart';
 
+part 'mock_system_date_time_format.dart';
 part 'mock_system_date_time_format_platform.dart';