+++ test/system_date_time_format/system_date_time_format_test.dart
@@ -1,5 +1,7 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
+import 'package:system_date_time_format/src/errors/sdtf_scope_not_found_error.dart';
 import 'package:system_date_time_format/src/system_date_time_format_platform_interface.dart';
 import 'package:system_date_time_format/system_date_time_format.dart';
 
@@ -8,16 +10,15 @@ import '../_tools/tools.dart';
 void main() {
   setUpAll(() {
     final platform = MockSystemDateTimeFormatPlatform();
-    String anyFallback() => any(named: 'fallback');
 
-    when(() => platform.getDateFormat(fallback: anyFallback()))
-        .thenAnswer((_) async => Stubs.dateFormat);
-    when(() => platform.getMediumDateFormat(fallback: anyFallback()))
-        .thenAnswer((_) async => Stubs.mediumDateFormat);
-    when(() => platform.getLongDateFormat(fallback: anyFallback()))
-        .thenAnswer((_) async => Stubs.longDateFormat);
-    when(() => platform.getTimeFormat(fallback: anyFallback()))
-        .thenAnswer((_) async => Stubs.timeFormat);
+    when(() => platform.getDatePattern())
+        .thenAnswer((_) async => Stubs.datePattern);
+    when(() => platform.getMediumDatePattern())
+        .thenAnswer((_) async => Stubs.mediumDatePattern);
+    when(() => platform.getLongDatePattern())
+        .thenAnswer((_) async => Stubs.longDatePattern);
+    when(() => platform.getTimePattern())
+        .thenAnswer((_) async => Stubs.timePattern);
 
     SystemDateTimeFormatPlatformInterface.instance = platform;
   });
@@ -30,26 +31,67 @@ void main() {
       expect(s1 == s2, true);
     });
 
-    test('dateFormat returns correct format: [${Stubs.dateFormat}]', () async {
-      await SystemDateTimeFormat().initialize();
-      expect(SystemDateTimeFormat().dateFormat, Stubs.dateFormat);
+    [
+      [SystemDateTimeFormat().getDatePattern, Stubs.datePattern],
+      [SystemDateTimeFormat().getMediumDatePattern, Stubs.mediumDatePattern],
+      [SystemDateTimeFormat().getLongDatePattern, Stubs.longDatePattern],
+      [SystemDateTimeFormat().getTimePattern, Stubs.timePattern],
+    ].forEach((input) {
+      final function = input.first as Future<String?> Function();
+      final expectedValue = input.second as String;
+
+      test('${function.name}() returns correct pattern: [$expectedValue]',
+          () async {
+        final result = await function();
+        expect(result, expectedValue);
+      });
     });
 
-    test('mediumDateFormat returns correct format: [${Stubs.mediumDateFormat}]',
+    test('getAllPatterns() returns correct patterns', () async {
+      final patterns = await SystemDateTimeFormat().getAllPatterns();
+      expect(patterns, Stubs.allPatterns);
+    });
+
+    testWidgets('of() throws SDTFScopeNotFoundError when there is NO SDTFScope',
+        (tester) async {
+      await tester.pumpWidget(const Placeholder());
+      final context = tester.element(find.byType(Placeholder));
+      expect(() => SystemDateTimeFormat.of(context),
+          throwsA(isA<SDTFScopeNotFoundError>()));
+    });
+
+    testWidgets('of() returns Patterns when there is SDTFScope',
+        (tester) async {
+      await tester.pumpWidget(const SDTFScope(child: Placeholder()));
+      await tester.pump();
+      final context = tester.element(find.byType(Placeholder));
+      final patterns = SystemDateTimeFormat.of(context);
+      expect(patterns, Stubs.allPatterns);
+    });
+
+    test('dateFormat returns correct pattern: [${Stubs.datePattern}]',
         () async {
       await SystemDateTimeFormat().initialize();
-      expect(SystemDateTimeFormat().mediumDateFormat, Stubs.mediumDateFormat);
+      expect(SystemDateTimeFormat().dateFormat, Stubs.datePattern);
     });
 
-    test('longDateFormat returns correct format: [${Stubs.longDateFormat}]',
+    test(
+        'mediumDateFormat returns correct pattern: [${Stubs.mediumDatePattern}]',
         () async {
       await SystemDateTimeFormat().initialize();
-      expect(SystemDateTimeFormat().longDateFormat, Stubs.longDateFormat);
+      expect(SystemDateTimeFormat().mediumDateFormat, Stubs.mediumDatePattern);
     });
 
-    test('timeFormat returns correct format: [${Stubs.timeFormat}]', () async {
+    test('longDateFormat returns correct pattern: [${Stubs.longDatePattern}]',
+        () async {
       await SystemDateTimeFormat().initialize();
-      expect(SystemDateTimeFormat().timeFormat, Stubs.timeFormat);
+      expect(SystemDateTimeFormat().longDateFormat, Stubs.longDatePattern);
+    });
+
+    test('timeFormat returns correct pattern: [${Stubs.timePattern}]',
+        () async {
+      await SystemDateTimeFormat().initialize();
+      expect(SystemDateTimeFormat().timeFormat, Stubs.timePattern);
     });
   });
 }